Add a test fixture to suppress the NuGet Preview Changes dialog
By default, this dialog is shown to users.  There is no Apex automation for this dialog currently, and all existing tests expect it not to show.  If it is not suppressed, the tests will hang.

For users who like to have this dialog enabled (guilty!), once suppressed, there is no UI affordance that I could find to restore it.

This change suppresses the dialog via the registry key, then restores it to it's previous value at the conclusion of the test.